requirements.txt
albumentations
torch
timm
torchvision
numpy
Pillow
pyyaml
python-dotenv
terratorch
segmentation-models-pytorch
